* temp_changes:
  Some more updates

Conflicts:
	Codechef/APRIL15/a.out
	Codechef/COOK57/a.exe
	Codechef/COOK58/a.out
	Codechef/COOK59/a.out
	Codechef/JULY15/MCHEF.cpp
	Codechef/JULY15/a.out
	Codechef/JULY15/data
	Codechef/JUNE15/a.out
	Codechef/LTIME22/a.out
	Codechef/LTIME23/a.out
	Codechef/LTIME24/a.out
	Codechef/MAY15/a.out
	Codechef/SNCK151A/a.out
	Codechef/SNCK151B/a.out
	Codechef/SNCK15EL/a.out
	Codeforces/Practice/a.out
	Codeforces/Practice/data
	Codeforces/Round_297/a.out
	Codeforces/Round_298/a.exe
	Codeforces/Round_299/a.out
	Codeforces/Round_301/a.out
	Codeforces/Round_304/a.out
	Codeforces/Round_305/a.out
	Codeforces/Round_306/a.out
	Codeforces/Round_308/a.out
	Codeforces/VKCup_2015/Round_1_Wildcard/picat
	Codeforces/ZeptoLab_CodeRush_15/a.out
	IOITC/2008/a.out
	IOITC/2013/a.out
	Kattis/a.out
	Kattis/data
	Kattis/worstweather.cpp
	Misc/Codejam/Qualification_2015/a.out
	Misc/Codejam/Round_1A_2015/a.out
	Misc/Codejam/Round_1B_2015/a.out
	Misc/IPSC/2015/B/a.out
	Misc/IPSC/2015/F/a.out
	Misc/IPSC/2015/G/a.out
	Misc/IPSC/2015/ipsc2015/j1.com
	Misc/IPSC/2015/ipsc2015/j2.com
	Misc/JOI/2015/a.out
	Misc/Yandex.Algorithm/Qualfication/a.out
	SPOJ/a.out
	SPOJ/data
	Topcoder/Round_1B_2015/a.out
	Topcoder/SRM 656/a.out
	Topcoder/SRM 658/a.out
	Topcoder/SRM 660/a.out
	USACO/US_Open_15/a.out

+++ Codechef/JULY15/MCHEF.cpp
@@ -21,7 +21,7 @@ void propagate(int i) {
   lazy[i] = INT_MAX/2;
 }
 
-void update(int seg_start, int seg_end, int i, int L, int R, long long val) {
+void update(int seg_start, int seg_end, int i, int L, int R, int val) {
   if(seg_start > seg_end) return;
   if(L > seg_end || R < seg_start) return;
 
@@ -41,7 +41,7 @@ void update(int seg_start, int seg_end, int i, int L, int R, long long val) {
   segtree[i] = min(segtree[2*i + 1], segtree[2*i + 2]);
 }
 
-long long query(int seg_start, int seg_end, int i, int L, int R) {
+int query(int seg_start, int seg_end, int i, int L, int R) {
   if(seg_start > seg_end) return INT_MAX/2;
   if(L > seg_end || R < seg_start) return INT_MAX/2;
 
@@ -60,7 +60,7 @@ int main() {
   int T; cin >> T;
 
   while(T--) {
-    long long N, K, M; cin >> N >> K >> M;
+    int N, K, M; cin >> N >> K >> M;
 
     for(int i = 0; i <= 4*(N + 1); i++) {
       segtree[i] = INT_MAX/2;
@@ -88,7 +88,7 @@ int main() {
       for(int j = 0; j <= K; j++) dp[0][j] = 0;
 
       for(int i = 1; i <= N; i++) {
-        for(long long j = 0; j <= K; j++) {
+        for(int j = 0; j <= K; j++) {
           if(min_cost[i] <= j) {
             dp[i][j] = max(dp[i - 1][j], dp[i - 1][j - min_cost[i]] + A[i]);
           }
@@ -101,7 +101,7 @@ int main() {
       total_sum += dp[N][K];
     }
     else {
-      vector<pair<long long, long long> > knapsack;
+      vector<pair<int, int> > knapsack;
       knapsack.push_back({0, 0});
       for(int i = 1; i <= N; i++) {
         if(A[i] > 0) knapsack.push_back({A[i], min_cost[i]});
@@ -112,7 +112,7 @@ int main() {
       for(int j = 0; j <= K; j++) dp[0][j] = 0;
 
       for(int i = 1; i <= K_N; i++) {
-        for(long long j = 0; j <= K; j++) {
+        for(int j = 0; j <= K; j++) {
           if(knapsack[i].second <= j) {
             dp[i][j] = max(dp[i - 1][j], dp[i - 1][j - knapsack[i].second] + knapsack[i].first);
           }
